+++ .gitignore
@@ -127,16 +127,21 @@ Dockerfile
 Makefile
 Makefile.*
 Settings.settings
+# Add exception for CMake support files
+!External/cmake/**/*.cmake
 
 # Ignore 3rd party libraries by default
 [Ll]ibraries/**
 [Ll]ib/**
 [Ll]ibs/**
 [Vv]endor/**
+.[Vv]env/
+.[Vv]env/**
 # Add exception for premake5 build system
 !Libraries/Premake5
 !Libraries/Premake5/premake5.exe
 !Libraries/Premake5/LICENSE.txt
+
 
 # Ignore all generated binaries by default
 [Bb]inaries/
