+++ External/cmake/cvars.cmake
@@ -1,0 +1,46 @@
+set(CVAR_VSYNC_ENABLED "gVsyncEnabled" CACHE STRING "")
+set(CVAR_Z_FIGHTING_MODE "gZFightingMode" CACHE STRING "")
+set(CVAR_NEW_FILE_DROPPED "gNewFileDropped" CACHE STRING "")
+set(CVAR_DROPPED_FILE "gDroppedFile" CACHE STRING "")
+set(CVAR_INTERNAL_RESOLUTION "gInternalResolution" CACHE STRING "")
+set(CVAR_MSAA_VALUE "gMSAAValue" CACHE STRING "")
+set(CVAR_SDL_WINDOWED_FULLSCREEN "gSdlWindowedFullscreen" CACHE STRING "")
+set(CVAR_TEXTURE_FILTER "gTextureFilter" CACHE STRING "")
+set(CVAR_IMGUI_CONTROLLER_NAV "gControlNav" CACHE STRING "")
+set(CVAR_CONSOLE_WINDOW_OPEN "gConsoleEnabled" CACHE STRING "")
+set(CVAR_CONTROLLER_CONFIGURATION_WINDOW_OPEN "gControllerConfigurationEnabled" CACHE STRING "")
+set(CVAR_CONTROLLER_DISCONNECTED_WINDOW_OPEN "gControllerDisconnectedWindowEnabled" CACHE STRING "")
+set(CVAR_CONTROLLER_REORDERING_WINDOW_OPEN "gControllerReorderingWindowEnabled" CACHE STRING "")
+set(CVAR_GFX_DEBUGGER_WINDOW_OPEN "gGfxDebuggerEnabled" CACHE STRING "")
+set(CVAR_STATS_WINDOW_OPEN "gStatsEnabled" CACHE STRING "")
+set(CVAR_ENABLE_MULTI_VIEWPORTS "gEnableMultiViewports" CACHE STRING "")
+set(CVAR_LOW_RES_MODE "gLowResMode" CACHE STRING "")
+set(CVAR_SIMULATED_INPUT_LAG "gSimulatedInputLag" CACHE STRING "")
+set(CVAR_GAME_OVERLAY_FONT "gOverlayFont" CACHE STRING "")
+set(CVAR_MENU_BAR_OPEN "gOpenMenuBar" CACHE STRING "")
+set(CVAR_PREFIX_CONTROLLERS "gControllers" CACHE STRING "")
+set(CVAR_PREFIX_ADVANCED_RESOLUTION "gAdvancedResolution" CACHE STRING "")
+add_compile_definitions(
+	CVAR_VSYNC_ENABLED="${CVAR_VSYNC_ENABLED}"
+	CVAR_Z_FIGHTING_MODE="${CVAR_Z_FIGHTING_MODE}"
+	CVAR_NEW_FILE_DROPPED="${CVAR_NEW_FILE_DROPPED}"
+	CVAR_DROPPED_FILE="${CVAR_DROPPED_FILE}"
+	CVAR_INTERNAL_RESOLUTION="${CVAR_INTERNAL_RESOLUTION}"
+	CVAR_MSAA_VALUE="${CVAR_MSAA_VALUE}"
+	CVAR_SDL_WINDOWED_FULLSCREEN="${CVAR_SDL_WINDOWED_FULLSCREEN}"
+	CVAR_TEXTURE_FILTER="${CVAR_TEXTURE_FILTER}"
+	CVAR_IMGUI_CONTROLLER_NAV="${CVAR_IMGUI_CONTROLLER_NAV}"
+	CVAR_CONSOLE_WINDOW_OPEN="${CVAR_CONSOLE_WINDOW_OPEN}"
+	CVAR_CONTROLLER_CONFIGURATION_WINDOW_OPEN="${CVAR_CONTROLLER_CONFIGURATION_WINDOW_OPEN}"
+	CVAR_CONTROLLER_DISCONNECTED_WINDOW_OPEN="${CVAR_CONTROLLER_DISCONNECTED_WINDOW_OPEN}"
+	CVAR_CONTROLLER_REORDERING_WINDOW_OPEN="${CVAR_CONTROLLER_REORDERING_WINDOW_OPEN}"
+	CVAR_GFX_DEBUGGER_WINDOW_OPEN="${CVAR_GFX_DEBUGGER_WINDOW_OPEN}"
+	CVAR_STATS_WINDOW_OPEN="${CVAR_STATS_WINDOW_OPEN}"
+	CVAR_ENABLE_MULTI_VIEWPORTS="${CVAR_ENABLE_MULTI_VIEWPORTS}"
+	CVAR_LOW_RES_MODE="${CVAR_LOW_RES_MODE}"
+	CVAR_SIMULATED_INPUT_LAG="${CVAR_SIMULATED_INPUT_LAG}"
+	CVAR_GAME_OVERLAY_FONT="${CVAR_GAME_OVERLAY_FONT}"
+	CVAR_MENU_BAR_OPEN="${CVAR_MENU_BAR_OPEN}"
+	CVAR_PREFIX_CONTROLLERS="${CVAR_PREFIX_CONTROLLERS}"
+	CVAR_PREFIX_ADVANCED_RESOLUTION="${CVAR_PREFIX_ADVANCED_RESOLUTION}"
+)
