--- Source/SampleApp/src/main.cpp
@@ -1,21 +1,0 @@
-#include <iostream>
-#include <vector>
-#include <string>
-
-#include "hello_statement.h"
-
-using namespace std;
-
-int main()
-{
-    vector<string> msg {"Hello", "C++", "World", "from", "VS Code", "and the C++ extension!"};
-
-    for (const string& word : msg)
-    {
-        cout << word << " ";
-    }
-    cout << endl;
-
-    PrintSecretMessage();
-
-}
