--- Source/SampleDLL/include/header_file.h
@@ -1,1 +1,0 @@
-#pragma once
