--- Source/SampleLibrary/include/Folder2/basic_addition.h
@@ -1,5 +1,0 @@
-#pragma once
-
-
-
-float AddTwoNumbers(float x, float y);
