--- Source/SampleLibrary/src/Folder1/secrets.cpp
@@ -1,13 +1,0 @@
-#include "../hello_statement.h"
-
-
-#define SECRET_MESSAGE "Using namespace std is bad practice!!"
-
-
-#include <iostream>
-#include <string>
-
-
-void PrintSecretMessage() {
-    std::cout << "Using namespace std is bad practice!!" << "\n";
-}
