--- Source/SampleLibrary/src/Folder2/basic_addition.cpp
@@ -1,7 +1,0 @@
-#include "basic_addition.h"
-
-
-
-float AddTwoNumbers(float x, float y) {
-    return x + y;
-}
