--- Source/SampleLibrary/src/counting.cpp
@@ -1,11 +1,0 @@
-#include "hello_statement.h"
-
-
-#include <vector>
-
-
-
-float CountToN(int N) {
-    std::vector<float> this_is_useless(N);
-    return 0.0f;
-}
