--- Source/SampleLibrary/src/hello_statement.h
@@ -1,9 +1,0 @@
-#pragma once
-
-
-
-// Defined in counting.cpp
-float CountToN(int N);
-
-// Defined in a secret location
-void PrintSecretMessage();
