+++ Source/libpotato/include/Folder2/basic_addition.h
@@ -1,0 +1,5 @@
+#pragma once
+
+
+
+float AddTwoNumbers(float x, float y);
