+++ Source/libpotato/include/hello_statement.h
@@ -1,0 +1,9 @@
+#pragma once
+
+
+
+// Defined in counting.cpp
+float CountToN(int N);
+
+// Defined in a secret location
+void PrintSecretMessage();
