+++ Source/libpotato/src/Folder2/basic_addition.cpp
@@ -1,0 +1,7 @@
+#include "basic_addition.h"
+
+
+
+float AddTwoNumbers(float x, float y) {
+    return x + y;
+}
