+++ Source/libpotato/src/sample.cpp
@@ -1,0 +1,1 @@
+#include "sample.h"
