+++ Source/libpotato/src/sample.h
@@ -1,0 +1,9 @@
+#pragma once
+
+
+
+namespace Potato {
+
+void SampleFunction();
+
+}
